* origin/master: (554 commits)
  SendReply: fix memory-free bug introduced in de70eb6 (fixes: 841)
  class library: make number of pings before server considered dead configurable
  Library: Wait for 5 /status failures before running Server actions
  Library/PmonoArtic: Fix the case of a rest as the first event
  Library: Avoid redundant releases for rests in PmonoArticStream
  documentation: document creating a new Dictionary from array of associations.
  class library: apply preprocessor to compiled files
  linux: install 2 more files needed in dev headers: SCVersion.txt, SC_PlugIn.hpp
  Change "global variable" references where "environment variables" are meant
  MP tutorial 10: Fix typos in an example (one causing syntax error)
  platform: standalone's modifyStartup needs to initialize openPorts
  Library: Psync's cleanup was incorrect and could yield a redundant rest
  documentation: improve thisFunction and thisFunctionDef
  streamline README.txt
  documentation: TempoClock: info on setting 'beats'
  documentation: update info on TempoClock (new use of logical time)
  sclang: TempoClock: use logical time instead of physical time
  sclang: allow changing TempoClock's beats
  sc class library: TempoClock: allow setting time base at creation
  README.txt: reworked and simplified with focus on SC IDE and version 3.6
  ...

Conflicts:
	QtCollider/CMakeLists.txt
	editors/sc-ide/CMakeLists.txt

+++ external_libraries/boost/boost/asio/detail/descriptor_ops.hpp
@@ -88,9 +88,9 @@ BOOST_ASIO_DECL bool non_blocking_write(int d,
 BOOST_ASIO_DECL int ioctl(int d, state_type& state, long cmd,
     ioctl_arg_type* arg, boost::system::error_code& ec);
 
-BOOST_ASIO_DECL int fcntl(int d, long cmd, boost::system::error_code& ec);
+BOOST_ASIO_DECL int fcntl(int d, int cmd, boost::system::error_code& ec);
 
-BOOST_ASIO_DECL int fcntl(int d, long cmd,
+BOOST_ASIO_DECL int fcntl(int d, int cmd,
     long arg, boost::system::error_code& ec);
 
 BOOST_ASIO_DECL int poll_read(int d,
